pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mode-watermarking"
version = "0.1.0"
description = "A watermarking system for Stable Diffusion models"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "thienphung00"},
]
keywords = ["watermarking", "stable-diffusion", "machine-learning", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

# Core ML/DL frameworks
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "diffusers>=0.21.0",
    "transformers>=4.30.0",
    # Numerical and image processing
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "scipy>=1.9.0",
    # Image quality metrics
    "lpips>=0.1.4",
    "scikit-image>=0.19.0",
    "pytorch-fid>=0.3.0",
    "torchmetrics>=0.11.0",
    # Machine learning utilities
    "scikit-learn>=1.2.0",
    # Visualization (optional, but needed for evaluation)
    "matplotlib>=3.5.0",
    "seaborn>=0.12.0",
    # Configuration and utilities
    "pyyaml>=6.0",
    "tqdm>=4.64.0",
    # Logging and experiment tracking (optional)
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    # Data download (optional, for COCO dataset)
    "kagglehub>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]

[project.urls]
Homepage = "https://github.com/thienphung00/mode-watermarking"
Repository = "https://github.com/thienphung00/mode-watermarking"
Issues = "https://github.com/thienphung00/mode-watermarking/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]

